+++ Source/Classes/ProcessValidateAttribute.ps1
@@ -4,16 +4,14 @@ class ProcessValidateAttribute : ValidateArgumentsAttribute {
    [void] Validate(
       [object] $arguments,
       [EngineIntrinsics] $EngineIntrinsics) {
-      
-      if (_hasProcessTemplateCacheExpired) { 
-         [VSTeamProcessCache]::processes = _getProcesses
-         [VSTeamProcessCache]::timestamp = (Get-Date).Minute
-      }
 
-      if (($null -ne [VSTeamProcessCache]::processes) -and (-not ($arguments -in [VSTeamProcessCache]::processes))) {
+      #Do not fail on null or empty, leave that to other validation conditions
+      if ([string]::IsNullOrEmpty($arguments)) {return}
+
+      if (($null -ne [VSTeamProcessCache]::GetCurrent()) -and (-not ($arguments -in [VSTeamProcessCache]::GetCurrent() )) ) {
          throw [ValidationMetadataException]::new(
             "'$arguments' is not a valid process. Valid processes are: '" +
-            ([VSTeamProcessCache]::processes -join "', '") + "'")
+            ([VSTeamProcessCache]::GetCurrent() -join "', '") + "'")
       }
    }
 }